reorganized team related screens
